# Display missing blocks in error message