@hinshun 2019-06-14
Curious why you think it has a significant performance implication?

The differences from what I see are:
- `if count != len(keys) {` vs `if set.Len() != 0 {`
  - `set.Len` is implemented by checking the length of a map.
- `count++` vs `set.Remove(nd.Cid())`
  - `set.Remove` is implemented by deleting a key from a map.