@Stebalien 2019-06-14
I'm worried about the error formatting. We could be missing hundreds of CIDs. Really, it probably won't be much of a performance issue (unless we end up storing these errors somewhere) but it'll definitely generate really long error messages.