@hinshun 2019-06-14
Okay that makes sense. In my case I still would like to know exactly which CIDs are missing in order to debug my application using IPFS, I'm perf testing various P2P scenarios, and one part of that is to shard a IPLD DAG across a IPFS cluster (i.e. node A gets ~1/3 of graph by size in bytes, node B gets another ~1/3, node C gets ~1/3). There was some bugs I had to work out in the sharding code and this made it difficult to diagnose what was going on.

Maybe we can clarify the error message `failed to fetch all nodes` by clarifying it failed to fetch `x` nodes to at least disambiguate it from network/etc errors. Is it possible to debug log it so the CIDs are still available?

I also wrote a small benchmark in case its interesting:
```sh
# master
❯ go test -run BenchmarkGetMany -bench=. -benchmem
goos: linux
goarch: amd64
pkg: github.com/ipfs/go-merkledag
BenchmarkGetMany1-8     	  100000	     10904 ns/op	     984 B/op	      20 allocs/op
BenchmarkGetMany10-8    	  100000	     11213 ns/op	     984 B/op	      20 allocs/op
BenchmarkGetMany100-8   	  100000	     10583 ns/op	     984 B/op	      20 allocs/op
PASS
ok  	github.com/ipfs/go-merkledag	6.397s

# hinshun:missing-nodes
❯ go test -run BenchmarkGetMany -bench=. -benchmem
goos: linux
goarch: amd64
pkg: github.com/ipfs/go-merkledag
BenchmarkGetMany1-8     	  100000	     11157 ns/op	    1176 B/op	      22 allocs/op
BenchmarkGetMany10-8    	  100000	     13112 ns/op	    1176 B/op	      22 allocs/op
BenchmarkGetMany100-8   	  100000	     12980 ns/op	    1176 B/op	      22 allocs/op
PASS
ok  	github.com/ipfs/go-merkledag	5.561s
```

```golang
func benchmarkGetMany(i int, b *testing.B) {
	ctx := context.Background()

	var nodes []ipld.Node
	for j := 0; j < i; j++ {
		nd := NodeWithData([]byte{byte(j)})
		nodes = append(nodes, nd)
	}

	var cids []cid.Cid
	for _, nd := range nodes {
		cids = append(cids, nd.Cid())
	}

	dagservs := make([]ipld.DAGService, b.N)
	for n := 0; n < b.N; n++ {
		srv := NewDAGService(dstest.Bserv())

		if err := srv.AddMany(ctx, nodes); err != nil {
			panic(err)
		}

		dagservs[n] = srv
	}

	b.ResetTimer()
	for n := 0; n < b.N; n++ {
		nds := dagservs[n].GetMany(ctx, cids)
		for _ = range nds {
		}
	}
}

func BenchmarkGetMany1(b *testing.B) { benchmarkGetMany(1, b) }

func BenchmarkGetMany10(b *testing.B) { benchmarkGetMany(1, b) }

func BenchmarkGetMany100(b *testing.B) { benchmarkGetMany(1, b) }
```