@Stebalien 2019-06-14
> benchmark

My performance concern was more about when we're missing blocks (drop the `AddMany` calls). In that case, the slow down looks like:

```
benchmark                 old ns/op     new ns/op     delta
BenchmarkGetMany1-4       7053          13330         +89.00%
BenchmarkGetMany10-4      8691          13055         +50.21%
BenchmarkGetMany100-4     8403          13370         +59.11%
```

However, this is, admittedly, a micro-benchmark. Really, let's just forget I said anything. This isn't an issue. The only real issues are:

1. The length of the error (ux).
2. The fact that the error won't actually include _all_ missing blocks.

> Okay that makes sense. In my case I still would like to know exactly which CIDs are missing

So, I suggested returning one CID because this won't return _all_ CIDs regardless. That is, `ipfs get` will traverse the graph, calling `GetMany` on the children of each node it finds. However, it'll stop when the first `GetMany` fails. That means it won't report missing nodes from other parts of the graph, just the children that are missing from the first block missing children.

> Maybe we can clarify the error message failed to fetch all nodes by clarifying it failed to fetch x nodes to at least disambiguate it from network/etc errors. Is it possible to debug log it so the CIDs are still available?

Yeah, that shouldn't be an issue (although, personally, I'd just insert print statements as required).