@hinshun 2019-06-14
Thanks for clarifying, I missed the fact that its only from the first block that failed to get all its children. I think ideally, the error message would look something like: `failed to fetch %d children of %s`, or if we want to incorporate one child: `failed to fetch all children of %s: %s (%d missing children)` where the one child is the first lexicographical sorted missing child?