@Stebalien 2019-06-15
> ideally, the error message would look something like:

Unfortunately, GetMany doesn't know that, it just knows that it's supposed to fetch a set of nodes. The part that's fetching children of blocks is in go-unixfs.